@@ -34,26 +34,27 @@ import de.cimt.talendcomp.xmldynamic.filter.TypeReadHandler;
 import de.cimt.talendcomp.xmldynamic.filter.WSDLSchemaFilter;
 import de.cimt.talendcomp.xmldynamic.filter.XMLFilterChain;
 import java.io.FileNotFoundException;
+import java.net.URISyntaxException;
 import java.text.DateFormat;
+import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Locale;
 import java.util.jar.JarInputStream;
 import java.util.jar.Manifest;
-import java.util.logging.Filter;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import javax.xml.parsers.ParserConfigurationException;
 import javax.xml.transform.TransformerException;
+import javax.xml.transform.TransformerFactoryConfigurationError;
 import org.xml.sax.SAXException;
 
 /**
  *
  * @author dkoch
  */
-public class XJCOptions extends Options {
-
+public class XJCOptions extends Options implements AutoCloseable{
     private static final Logger LOG = Logger.getLogger("de.cimt.talendcomp.xmldynamic");
     public boolean extendClasspath = true;
     public boolean compileSource = true;
@@ -65,36 +66,39 @@ public class XJCOptions extends Options {
     public boolean forceGenerate = false;
     public boolean printGrammar = false;
     public boolean createJar = false;
-    public String  jarFilePath = null;
-    public String  checksumValue = "";
-    public String  targetName = "gen_" + Util.uniqueString() + ".jar";
-    public String  grammarFilePath = null;
+    public String jarFilePath = null;
+    public String checksumValue = "";
+    public String targetName = "gen_" + Util.uniqueString() + ".jar";
+    public String grammarFilePath = null;
     public long newestGrammar = 0l;
     public static final String VERSION;
     public static final String LASTUPDATE;
-    
+
     static {
-        String versionString="unknown";
-        String d="";
+        String versionString = "latest";
+        String d = "";
         try {
-            JarInputStream jis=new JarInputStream(XJCOptions.class.getProtectionDomain().getCodeSource().getLocation().openStream());
+            JarInputStream jis = new JarInputStream(XJCOptions.class.getProtectionDomain().getCodeSource().getLocation().openStream());
             final Manifest manifest = jis.getManifest();
             try {
-                Date date=new SimpleDateFormat("yyyyMMdd-HHmm").parse( manifest.getMainAttributes().getValue("Implementation-Timestamp") );
-                d=DateFormat.getDateInstance(2, Locale.getDefault()).format(date);
-            } catch (Throwable t) {}
-            versionString=manifest.getMainAttributes().getValue("Implementation-Version");
-        } catch (Throwable t) {}
-        VERSION   =versionString;
-        LASTUPDATE=d;
+                Date date = new SimpleDateFormat("yyyyMMdd-HHmm").parse(manifest.getMainAttributes().getValue("Implementation-Timestamp"));
+                d = DateFormat.getDateInstance(2, Locale.getDefault()).format(date);
+            } catch (NullPointerException | ParseException t) {
+            }
+            versionString = manifest.getMainAttributes().getValue("Implementation-Version");
+        } catch (  Throwable t) {
+            t.printStackTrace();
+        }
+        VERSION = versionString;
+        LASTUPDATE = d;
     }
-    
+
     // stores relations between source and alias
-    private final Map<String, String> grammarCache = new HashMap<String, String>();
-    
+    private final Map<String, String> grammarCache = new HashMap<>();
+
     // temporary directory used to store modified grammars
     private final File tmproot;
-    private final List<Pair<String, String>> _complexTypes = new ArrayList<Pair<String, String>>();
+    private final List<Pair<String, String>> _complexTypes = new ArrayList<>();
     private final Map<Pair<String, String>, AtomicInteger> _usageCount = new HashMap<Pair<String, String>, AtomicInteger>() {
         private static final long serialVersionUID = 1L;
 
@@ -109,9 +113,10 @@ public class XJCOptions extends Options {
             return val;
         }
     };
-    
+
     /**
-     * used to activate printing of manipulated grammars before generating code model
+     * used to activate printing of manipulated grammars before generating code
+     * model
      */
     public XJCOptions() {
         super();
@@ -119,12 +124,12 @@ public class XJCOptions extends Options {
         try {
             tmpfile = File.createTempFile("2890374092", "092830198");
         } catch (IOException ex) {
-            LOG.log(Level.SEVERE,"Error creating temporary file", ex);
+            LOG.log(Level.SEVERE, "Error creating temporary file", ex);
             throw new RuntimeException("temp not available");
         }
         tmproot = new File(tmpfile.getParentFile(), Util.uniqueString());
         if (LOG.isLoggable(Level.FINER)) {
-            LOG.finer("Set tmp-root to: "+tmproot.getAbsolutePath());
+            LOG.log(Level.FINER, "Set tmp-root to: {0}", tmproot.getAbsolutePath());
         }
         tmproot.mkdirs();
         tmproot.deleteOnExit();
@@ -135,7 +140,7 @@ public class XJCOptions extends Options {
         enableIntrospection = true;
         verbose = true;
     }
-    
+
     private Set<Pair<String, String>> getComplexTypes() {
         return _usageCount.entrySet()
                 .stream()
@@ -144,24 +149,23 @@ public class XJCOptions extends Options {
                 .map(p -> p.getKey())
                 .collect(Collectors.toSet());
     }
-    
+
     @Override
-    protected void finalize() throws Throwable {
+    public void close() {
         for (File f : tmproot.listFiles()) {
             f.delete();
         }
         tmproot.delete();
-        super.finalize();
     }
 
     @Override
     public synchronized void addGrammar(final InputSource source) throws RuntimeException {
         /**
-         * Changed behavior as this method never send the original inputsource 
-         * to the superclass. 
-         * It uses a set of filters to pre-process the original input and to parse 
-         * it as one or multiple expands sources to a in-memory-source instance.
-         * this one will be transferred to the superclass.
+         * Changed behavior as this method never send the original inputsource
+         * to the superclass. It uses a set of filters to pre-process the
+         * original input and to parse it as one or multiple expands sources to
+         * a in-memory-source instance. this one will be transferred to the
+         * superclass.
          */
         try {
 
@@ -181,7 +185,7 @@ public class XJCOptions extends Options {
                     @Override
                     public synchronized String createInputSource(String xmlbuffer) {
                         final String id = Util.uniqueString() + ".xsd";
-                        addGrammar( new InMemorySource(xmlbuffer, id) );
+                        addGrammar(new InMemorySource(xmlbuffer, id));
                         return id;
                     }
 
@@ -190,9 +194,9 @@ public class XJCOptions extends Options {
             }
             if (source.getClass().equals(InMemorySource.class)) {
                 alias = ((InMemorySource) source).alias;
-            } 
-            
-            chain.add( new TypeReadHandler(){
+            }
+
+            chain.add(new TypeReadHandler() {
                 @Override
                 public int incrementUsageCount(Pair<String, String> type) {
                     return _usageCount.get(type).incrementAndGet();
@@ -203,9 +207,9 @@ public class XJCOptions extends Options {
                     _complexTypes.add(complexType);
                 }
             });
-            
+
             DependencyFilter df = new DependencyFilter(rootURI) {
-                
+
                 @Override
                 protected synchronized String getRelocatedSchemaLocation(URI root, String location) {
                     if (location == null || location.length() == 0) {
@@ -216,13 +220,13 @@ public class XJCOptions extends Options {
 
                         if (!nestedUri.isAbsolute()) {
                             // memory sources are already analyzed...
-                            if (root.getScheme().equalsIgnoreCase("mem") && grammarCache.containsValue(location) ) {
+                            if (root.getScheme().equalsIgnoreCase("mem") && grammarCache.containsValue(location)) {
                                 return location;
                             }
                             nestedUri = root.resolve(nestedUri);
                         }
                         String systemID = nestedUri.toString();
-                                
+
                         // only not handled sources must be analyzed...
                         if (!grammarCache.containsKey(systemID)) {
                             InputSource source = new InputSource(systemID);
@@ -230,7 +234,7 @@ public class XJCOptions extends Options {
                             addGrammar(source);
                         }
                         return grammarCache.get(systemID);
-                    } catch (Exception ex) {
+                    } catch (RuntimeException | URISyntaxException ex) {
                         throw new RuntimeException(ex);
                     }
                 }
@@ -249,15 +253,15 @@ public class XJCOptions extends Options {
             StringWriter w = new StringWriter();
             TransformerFactory.newInstance().newTransformer().transform(new SAXSource(reader, source), new StreamResult(w));
 
-            InMemorySource ims = new InMemorySource(w.toString(), alias) ;
+            InMemorySource ims = new InMemorySource(w.toString(), alias);
             if (ims.isEmtpy()) {
                 return;
             }
             grammarCache.put(rootURI.toString(), alias);
 
-            super.addGrammar( new InMemorySource(w.toString(), alias) );
-        } catch (Throwable ex) {
-        	LOG.log(Level.SEVERE, Messages.format("PRELOAD.ANALYSE.FAILED"), ex);
+            super.addGrammar(new InMemorySource(w.toString(), alias));
+        } catch (ParserConfigurationException | TransformerException | TransformerFactoryConfigurationError | SAXException ex) {
+            LOG.log(Level.SEVERE, Messages.format("PRELOAD.ANALYSE.FAILED"), ex);
             throw new RuntimeException(Messages.format("PRELOAD.ANALYSE.FAILED"), ex);
         }
     }
@@ -272,7 +276,7 @@ public class XJCOptions extends Options {
             SAXParserFactory spf = SAXParserFactory.newInstance();
             spf.setNamespaceAware(true);
             spf.setXIncludeAware(true);
-            
+
             final Set<Pair<String, String>> typeBindings = getComplexTypes();
             PluginFilter pluginFilter = new PluginFilter() {
                 @Override
@@ -286,34 +290,34 @@ public class XJCOptions extends Options {
             if (createGraph) {
                 chain.add(new GraphFilter());
             }
-            
+
             if (ignoreAnnotations) {
-            	chain.add(new AnnotationFilter());
+                chain.add(new AnnotationFilter());
             }
             chain.add(pluginFilter);
             ChecksumFilter csf = new ChecksumFilter();
             chain.add(csf);
 
-            List<InputSource> ng = new ArrayList<InputSource>();
+            List<InputSource> ng = new ArrayList<>();
             XMLFilteredReader reader = new XMLFilteredReader(spf.newSAXParser().getXMLReader(), chain);
             final Transformer transformer = TransformerFactory.newInstance().newTransformer();
             for (InputSource source : super.getGrammars()) {
                 File res = new File(tmproot, ((InMemorySource) source).alias);
                 if (LOG.isLoggable(Level.FINER)) {
-                	LOG.finer("Use as temporary xsd result file: " + res.getAbsolutePath());
+                    LOG.log(Level.FINER, "Use as temporary xsd result file: {0}", res.getAbsolutePath());
                 }
                 transformer.transform(
-                    new SAXSource(reader, source),
-                    new StreamResult(res)
+                        new SAXSource(reader, source),
+                        new StreamResult(res)
                 );
-                
+
                 InputSource exportedGrammar = new InputSource(new FileInputStream(res));
-                exportedGrammar.setSystemId( res.toURI().toString() );
-                ng.add( exportedGrammar );
+                exportedGrammar.setSystemId(res.toURI().toString());
+                ng.add(exportedGrammar);
             }
-            
+
             checksumValue = csf.toString();
-            return ng.toArray( new InputSource[ng.size()] );
+            return ng.toArray(InputSource[]::new);
         } catch (FileNotFoundException | ParserConfigurationException | TransformerException | SAXException ex) {
             LOG.log(Level.SEVERE, "Error transforming grammar", ex);
             throw new RuntimeException(ex);
@@ -322,7 +326,8 @@ public class XJCOptions extends Options {
 
     /**
      * Recursively scan directories and add all XSD files in it.
-     * @param dir folder or file 
+     *
+     * @param dir folder or file
      */
     @Override
     public void addGrammarRecursive(File dir) {
@@ -350,8 +355,8 @@ public class XJCOptions extends Options {
             return bindfiles;
         }
         /**
-         * when enableBasicSubstitution is set, activate replacement for xsd:duration, 
-         * xsd:date, xsd:dateTime and enable plugin xjc-simple
+         * when enableBasicSubstitution is set, activate replacement for
+         * xsd:duration, xsd:date, xsd:dateTime and enable plugin xjc-simple
          */
         InputSource bind = new InputSource(XJCOptions.class.getResourceAsStream("bindings.xml"));
         bind.setSystemId(XJCOptions.class.getResource("bindings.xml").toString());
@@ -363,16 +368,15 @@ public class XJCOptions extends Options {
 
     @Override
     public void addGrammar(File source) {
-            if (source == null) {
-                    throw new IllegalArgumentException("source file must not be null");
-            }
-            this.grammarFilePath = source.getAbsolutePath();
-            super.addGrammar(source);
+        if (source == null) {
+            throw new IllegalArgumentException("source file must not be null");
+        }
+        this.grammarFilePath = source.getAbsolutePath();
+        super.addGrammar(source);
 
-            if(source.lastModified()>newestGrammar){
-                newestGrammar=source.lastModified();
-            }
+        if (source.lastModified() > newestGrammar) {
+            newestGrammar = source.lastModified();
+        }
     }
 
-    
 }
